Enable create and read data from firebase
