+++ src/pages/Main.jsx
@@ -3,23 +3,9 @@ import {Button, Form, Modal} from "react-bootstrap";
 import Nav from "../components/Nav";
 import TILCard from '../components/TILCard';
 import Footer from "../components/Footer";
-import axios from "axios";
 
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_API_KEY,
-  authDomain: process.env.REACT_APP_AUTH_DOMAIN,
-  projectId: process.env.REACT_APP_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_APP_ID,
-  measurementId: process.env.REACT_APP_MEASUREMENT_ID
-}
-
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
+import { dbService } from "../fbase";
+import { addDoc, getDocs, query, collection } from "firebase/firestore";
 
 const Main = () => {
   const [cardData, setCardData] = useState([]);
@@ -27,20 +13,26 @@ const Main = () => {
   const showModal = () => setIsModal(true);
   const hideModal = () => setIsModal(false);
 
+  const padTime = (time) => time.toString().length <= 1 ? '0' + time : time;
+
   const getData = async () => {
-    const res = await axios.get("http://localhost:4000/posts?_sort=id&_order=desc");
-    setCardData(res.data);
+    setCardData([]);
+    const res = await getDocs(query(collection(dbService, "posts")));
+    // res.forEach(elem => setCardData([...cardData, elem.data()]));
+    res.forEach(elem => setCardData(prevData => [...prevData, elem.data()]));
   }
 
   const postData = async (e) => {
     const date = new Date();
     e.preventDefault();
-    await axios.post("http://localhost:4000/posts", {
+
+    await addDoc(collection(dbService, "posts"), {
+      "key": Date.now(),
       "id": Date.now(), 
       "title": e.target[0].value,
       "body": e.target[1].value,
-      "published_at": `${date.getFullYear()}-${date.getMonth()+1}-${date.getDate()+1}`
-    })
+      "published_at": `${date.getFullYear()}-${padTime(date.getMonth()+1)}-${padTime(date.getDate()+1)}`
+    });
     hideModal();
     getData();
   }
@@ -55,7 +47,7 @@ const Main = () => {
     {
       cardData.map(card => 
         <TILCard
-          key={card.id} 
+          key={card.key}
           id={card.id}
           title={card.title}
           body={card.body}
